refactor: extract 'Context.unregisterReceiverSilently' to :common:android
Removes a dependency on com.ichi2.anki.utils.ext from widget config
activities, helping break circular dependencies for feature module
extraction
